% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{as.data.frame.NONMEMdata}
\alias{as.data.frame.NONMEMdata}
\title{Convert a NONMEMdata object to a data frame}
\usage{
\method{as.data.frame}{NONMEMdata}(x, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{x}{The object to convert to a data.frame}

\item{..., stringsAsFactors}{Ignored (kept as arguments for
compatibility with the generic as.data.frame method)}
}
\value{
A data.frame (or similar object) as described in the details
}
\description{
Convert a NONMEMdata object to a data frame
}
\details{
The data.frame will be sorted by ID, TIME (NA first then
  ordered by number), EVID (decreasing), CMT, then the group columns.
  Columns will be converted to numeric with the following rules:
  
\itemize{
\item{numeric}{No conversion}
\item{logical}{as.numeric (0 or 1)}
\item{character}{No conversion, except for the ID column.  The ID
column will be converted to a factor if it is not already numeric.}
\item{factor}{as.numeric on the factor with a new name for the factor
column.  If the factor column is mapped to a NONMEM column (often 
used for ID), then the numeric column will be named with the NONMEM 
mapping.  If the factor column is not named, the numbered column will
have "_numeric" appended to the original column name, the
factor will be converted to a character column}
\item{any other class}{Coerce to character with as.character}
}

Column order will be ID, TIME, EVID, EXCLUDENUM, other numeric
columns (after conversion), character columns (after conversion).
}
