% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{as.NONMEMdata}
\alias{as.NONMEMdata}
\alias{as.NONMEMdata.PKNCAdata}
\alias{as.NONMEMdata.PKNCAconc}
\alias{as.NONMEMdata.PKNCAdose}
\title{Convert a PKNCA object into a NONMEM-ready dataset}
\usage{
as.NONMEMdata(object, ...)

\method{as.NONMEMdata}{PKNCAdata}(object, ..., conc_cmt_map, dose_cmt_map)

\method{as.NONMEMdata}{PKNCAconc}(object, ..., conc_cmt_map = NULL)

\method{as.NONMEMdata}{PKNCAdose}(object, ..., dose_cmt_map = NULL)
}
\arguments{
\item{object}{The object to make into a NONMEM-ready dataset.  To 
ensure consistent output of reset data, it's preferred to provide a
PKNCAdata object.}

\item{...}{Ignored}

\item{conc_cmt_map, dose_cmt_map}{A data.frame (or similar object) 
that indicates the mapping between concentration or dosing data and
compartment names.  See details for more information.}
}
\value{
A data.frame or similar ready for export to a .csv file as a 
  NONMEM-ready dataset with an attribute "name_map" for mapping the 
  data.frame names to NONMEM names.
}
\description{
Convert a PKNCA object into a NONMEM-ready dataset
}
\details{
The attribute "NONMEM_column" will be added to columns to
  define the following.
  
  For every data type:
  
  \itemize{
   \item{ID}{the PKNCA subject}
   \item{TIME}{the time (right hand side) column}
   \item{EVID}{0 for PKNCAconc, 1 for PKNCAdose, 3 for reset}
   \item{EXCLUDETEXT}{the text of the exclude column; this will be
     converted to a factor and put into the EXCLUDENUM column on
     conversion to a data.frame (and writing to a file)}
  }
  
  For concentration data, name mapping for every data type plus:
  
  \itemize{
   \item{DV}{concentration measurement (or generally, the independent variable for PKNCAconc}
  }
  
  For dose data, name mapping for every data type plus:
  
  \itemize{
   \item{AMT}{amount of the dose for PKNCAdose}
   \item{RATE}{the rate of dosing for PKNCAdose}
   \item{II}{interdose interval is set to 0}
   \item{ADDL}{additional dose count is set to 0}
  }
  
  \code{conc_cmt_map} and \code{dose_cmt_map} will use the 
  \code{dplyr::left_join} function to combine with the data, so 
  matching columns will define the mapping.  One column must be named
  "CMT" in the data.frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{PKNCAdata}: Convert a PKNCAdata object to NONMEM data

\item \code{PKNCAconc}: Convert a PKNCAconc object to NONMEM data

\item \code{PKNCAdose}: Convert a PKNCAdose object to NONMEM data
}}

