% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SDTM.R
\name{as.PKNCAconc.SDTM_PC}
\alias{as.PKNCAconc.SDTM_PC}
\title{Convert an SDTM_PC object to a PKNCAconc object}
\usage{
as.PKNCAconc.SDTM_PC(x, ..., fun_time_nominal = sdtm_time_nominal,
  fun_time_actual = NULL)
}
\arguments{
\item{x}{The SDTM object to convert}

\item{...}{arguments passed to PKNCAconc}

\item{fun_time_nominal}{The function to calculate the nominal times.}

\item{fun_time_actual}{The function to calculate the actual times.}
}
\value{
A PKNCAconc object
}
\description{
Convert an SDTM_PC object to a PKNCAconc object
}
\seealso{
\code{\link{as.PKNCAdose.SDTM_EX}}
}
