% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SDTM.R
\name{as.SDTM}
\alias{as.SDTM}
\alias{as.SDTM.data.frame}
\alias{as.SDTM.tbl_df}
\title{Convert an object into an SDTM object}
\usage{
as.SDTM(x, ...)

\method{as.SDTM}{data.frame}(x, ...)

\method{as.SDTM}{tbl_df}(x, ...)
}
\arguments{
\item{x}{The object to convert}
}
\value{
An object with the same class and classes of the SDTM domain
  (like "SDTM_PC") and "SDTM" classes added.
}
\description{
Convert an object into an SDTM object
}
