% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messy_groups.R
\name{clean_messy_groups}
\alias{clean_messy_groups}
\alias{replace_messy_groups}
\title{Simplify data to a single value within a group}
\usage{
clean_messy_groups(data, keys, values, missing_value = NA)

replace_messy_groups(data, keys, values, missing_value = NA)
}
\arguments{
\item{data}{The dataset (typically a data.frame or similar) to
simplify}

\item{keys}{Column names (as character strings) for grouping}

\item{values}{Columns to from the original data as values that must
be unique within a group (defaults to all columns other than
\code{keys})}

\item{missing_value}{A scalar value indicating what should be 
considered missing and should be used for replacement if all values
are missing (typically \code{NA}).}
}
\value{
A dataset with one row per key and one value in each column.
}
\description{
During dataset generation, sometimes data from different sources may 
put missing values where consistent results should be present.
}
\section{Functions}{
\itemize{
\item \code{replace_messy_groups}: Both clean the groups and replace them
back in the original dataset.
}}

