% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SDTM.R
\name{detect_SDTM}
\alias{detect_SDTM}
\title{Detect the SDTM domain for an object and add that class to the class
of the object.}
\usage{
detect_SDTM(x)
}
\arguments{
\item{x}{The object to detect using the "DOMAIN" column in the data}
}
\value{
The same object with the SDTM domain and SDTM classes added.
}
\description{
Detect the SDTM domain for an object and add that class to the class
of the object.
}
