% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{get_NONMEM_name_map}
\alias{get_NONMEM_name_map}
\title{Extract the NONMEM-mapped columns from the names of the data}
\usage{
get_NONMEM_name_map(object)
}
\arguments{
\item{object}{The object that may have NONMEM-named columns}
}
\value{
A named vector.  Names are the NONMEM names; values are the
  data column names.
}
\description{
Extract the NONMEM-mapped columns from the names of the data
}
\seealso{
\code{\link{set_NONMEM_name_map}}
}
