% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{names<-.NONMEMdata}
\alias{names<-.NONMEMdata}
\title{Rename NONMEMdata columns ensuring that name_map and groups are also
updated.}
\usage{
\method{names}{NONMEMdata}(x) <- value
}
\arguments{
\item{x}{A NONMEMdata object}

\item{value}{The new name vector}
}
\value{
An updated NONMEMdata object
}
\description{
Rename NONMEMdata columns ensuring that name_map and groups are also
updated.
}
