% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserve_attributes.R
\name{preserve_attributes}
\alias{preserve_attributes}
\alias{preserve_attributes<-}
\title{Return a subset of the attributes as a list}
\usage{
preserve_attributes(x, which)

preserve_attributes(x) <- value
}
\arguments{
\item{x}{The object to extract attributes from}

\item{which}{A character string of attributes to preserve}
}
\value{
A list of attributes
}
\description{
Return a subset of the attributes as a list
}
\section{Functions}{
\itemize{
\item \code{preserve_attributes<-}: Assign a subset of attributes to an object
}}

