% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.if.required.R
\name{quote.if.required}
\alias{quote.if.required}
\title{Update a vector with the minimum quotes required to make a .csv file}
\usage{
quote.if.required(x, qmethod = c("double", "escape"), sep = ",",
  eol = "\\n")
}
\arguments{
\item{x}{The vector (either character or something that can be
coerced into a character vector)}

\item{qmethod, sep, eol}{See \code{write.csv}}
}
\value{
A vector as a character string, quoted if required
}
\description{
Update a vector with the minimum quotes required to make a .csv file
}
