% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{rbind.NONMEMdata}
\alias{rbind.NONMEMdata}
\title{Combine multiple NONMEMdata tables while matching/maintaining the
"name_map" and "groups" attributes}
\usage{
\method{rbind}{NONMEMdata}(..., deparse.level = 0)
}
\arguments{
\item{...}{values to combine}

\item{deparse.level}{Ignored}
}
\value{
A new NONMEMdata object with the combined name_maps from the
  source.
}
\description{
Combine multiple NONMEMdata tables while matching/maintaining the
"name_map" and "groups" attributes
}
