% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.SDTM.R
\name{sdtm_time_nominal}
\alias{sdtm_time_nominal}
\alias{sdtm_time_start_nominal}
\alias{sdtm_time_end_nominal}
\title{Calculate the nominal time from SDTM data}
\usage{
sdtm_time_nominal(x, col_day_suffix = "DY", col_time_suffix = "TPTNUM")

sdtm_time_start_nominal(x, col_day_suffix = "STDY",
  col_time_suffix = "TPTNUM")

sdtm_time_end_nominal(x, col_day_suffix = "ENDY",
  col_time_suffix = "TPTNUM")
}
\arguments{
\item{x}{The object to extract the nominal time from}

\item{col_day_suffix}{The suffix (after the domain) for the day
column (e.g. "PCDY")}

\item{col_time_suffix}{The suffix (after the domain) for the nominal
time numeric column}
}
\value{
The vector of numeric times.
}
\description{
Calculate the nominal time from SDTM data
}
\section{Functions}{
\itemize{
\item \code{sdtm_time_start_nominal}: Calculate the nominal start time

\item \code{sdtm_time_end_nominal}: Calculate the nominal end time
}}

