% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{set_NONMEM_name_map}
\alias{set_NONMEM_name_map}
\title{Extract the NONMEM-mapped columns from the names of the data}
\usage{
set_NONMEM_name_map(object, name_map, reset = TRUE)
}
\arguments{
\item{object}{The object that may have NONMEM-named columns}

\item{name_map}{A named vector.  Names are the NONMEM names; values
are the data column names.}

\item{reset}{For any columns not named in \code{name_map}, unset the
"NONMEM_column" attribute.}
}
\value{
The object with columns updated to have attributes named
  "NONMEM_column" with a value of the NONMEM column name.
}
\description{
Extract the NONMEM-mapped columns from the names of the data
}
\seealso{
\code{\link{get_NONMEM_name_map}}
}
