% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{switchNames}
\alias{switchNames}
\title{Switch the names and values in a vector.}
\usage{
switchNames(x)
}
\arguments{
\item{x}{The object to switch}
}
\value{
x with the names put into the values and the values put into
  the names.
}
\description{
Switch the names and values in a vector.
}
\seealso{
\code{\link{setNames}}
}
