% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{write.NONMEMdata}
\alias{write.NONMEMdata}
\title{Write a NONMEMdata object to a .csv file ready for analysis}
\usage{
write.NONMEMdata(x, file, also_RDS = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{A NONMEMdata object}

\item{file}{The file to write to (see \code{write.csv} for details)}

\item{also_RDS}{Also output an .RDS file with ".RDS" appended to the
filename}

\item{verbose}{Passed to as.data.frame.NONMEMdata}
}
\value{
x (invisibly)
}
\description{
Write a NONMEMdata object to a .csv file ready for analysis
}
\details{
The output will be sorted based on the ID column, the TIME 
  column, EVID (decreasing), CMT, then the group columns.
}
