% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.NONMEM.R
\name{as.NONMEMreset}
\alias{as.NONMEMreset}
\alias{as.NONMEMreset.default}
\alias{as.NONMEMreset.NULL}
\title{Generate one NONMEM reset record for each grouping in a data set.}
\usage{
as.NONMEMreset(object, ...)

\method{as.NONMEMreset}{default}(object, ...)

\method{as.NONMEMreset}{NULL}(object, ...)
}
\arguments{
\item{object}{The object to use as the source for the reset record.}
}
\value{
A data.frame or similar object ready for 

  \itemize{
    \item{ID}{the PKNCA subject}
    \item{TIME}{\code{NA_real_}}
    \item{EVID}{3 for all records (indicating reset)}
    \item{EXCLUDETEXT}{\code{NA_character_}}
  }
}
\description{
Generate one NONMEM reset record for each grouping in a data set.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Generate reset records for objects that
can be made into NONMEMdata objects

\item \code{NULL}: Returns \code{NULL} (typically in case
dose is not present)
}}

